src/app/features/pdf-tools/convert-to-pdf/word-to-pdf/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-docx==1.1.0
reportlab==4.1.0 